requirements.txt
python-telegram-bot>=20.0
selenium>=4.10.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
webdriver_manager>=4.0.0
pytest>=7.4.0
logging>=0.5.1.2